update + adding bonnus functions

+++ ft_strcat.c
@@ -6,7 +6,7 @@
 /*   By: jfortin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 10:24:07 by jfortin           #+#    #+#             */
-/*   Updated: 2015/11/24 10:24:08 by jfortin          ###   ########.fr       */
+/*   Updated: 2015/11/26 14:42:41 by jfortin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strcat(char *dest, const char *src)
 {
-	unsigned int	i;
-	size_t			dest_len;
+	size_t	i;
+	size_t	dest_len;
 
 	i = 0;
 	dest_len = ft_strlen(dest);
